+++ app/scripts/services/layer-group-service.js
@@ -420,7 +420,8 @@ angular.module('lizard-nxt')
           type: layer.type,
           layerGroupSlug: lgSlug,
           layerSlug: layer.slug,
-          aggType: layer.aggregation_type
+          aggType: layer.aggregation_type,
+          summary: layer.summary
         });
       };